+++ packages/brick-kit/src/bindListeners.spec.ts
@@ -342,6 +342,7 @@ describe("bindListeners", () => {
         {
           action: "message.subscribe",
           args: [
+            "task1",
             {
               system: "pipeline",
               topic: "pipeline.running.001",
@@ -359,6 +360,7 @@ describe("bindListeners", () => {
         {
           action: "message.unsubscribe",
           args: [
+            "task1",
             {
               system: "pipeline",
               topic: "pipeline.running.001",
@@ -387,11 +389,17 @@ describe("bindListeners", () => {
           target: "#target-elem",
           method: "forAsyncWillSuccess",
           callback: {
-            success: {
-              action: "console.log",
-            },
+            success: [
+              {
+                action: "console.log",
+              },
+              {
+                action: "console.log",
+                args: ["<% EVENT.oops() %>"],
+              },
+            ],
             error: {
-              action: "console.error",
+              action: "console.warn",
             },
           },
         },
@@ -403,7 +411,7 @@ describe("bindListeners", () => {
               action: "console.log",
             },
             error: {
-              action: "console.error",
+              action: "console.warn",
             },
             finally: {
               action: "console.info",
@@ -528,12 +536,29 @@ describe("bindListeners", () => {
       "callback.success"
     );
     expect((console.log as jest.Mock).mock.calls[2][0].detail).toBe("resolved");
+
     expect(console.info).toBeCalledTimes(2);
     expect(console.info).toBeCalledWith(event1);
     expect((console.info as jest.Mock).mock.calls[1][0].type).toBe(
       "callback.finally"
     );
-    expect(console.warn).toBeCalledWith("specified args for console.warn");
+
+    expect(console.warn).toBeCalledTimes(3);
+    expect(console.warn).toHaveBeenNthCalledWith(
+      1,
+      "specified args for console.warn"
+    );
+    expect(console.warn).toHaveBeenNthCalledWith(
+      2,
+      expect.stringContaining("context.assign")
+    );
+    expect(console.warn).toHaveBeenNthCalledWith(
+      3,
+      new CustomEvent("callback.error", {
+        detail: "oops",
+      })
+    );
+
     expect(console.error).toBeCalledTimes(7);
     expect(console.error).toHaveBeenNthCalledWith(
       1,
@@ -554,10 +579,12 @@ describe("bindListeners", () => {
       6,
       'Error: Provider not defined: "not-defined-provider".'
     );
-    expect((console.error as jest.Mock).mock.calls[6][0].type).toBe(
-      "callback.error"
+    expect(console.error).toHaveBeenNthCalledWith(
+      7,
+      expect.objectContaining({
+        message: expect.stringContaining("EVENT.oops is not a function"),
+      })
     );
-    expect((console.error as jest.Mock).mock.calls[6][0].detail).toBe("oops");
 
     expect((sourceElem as any).forGood).toHaveBeenNthCalledWith(
       1,
@@ -588,6 +615,7 @@ describe("bindListeners", () => {
     });
 
     expect(mockMessageDispatcher.subscribe).toHaveBeenLastCalledWith(
+      "task1",
       { system: "pipeline", topic: "pipeline.running.001" },
       expect.objectContaining({
         brick: sourceElem,
@@ -600,6 +628,7 @@ describe("bindListeners", () => {
       })
     );
     expect(mockMessageDispatcher.unsubscribe).toHaveBeenLastCalledWith(
+      "task1",
       { system: "pipeline", topic: "pipeline.running.001" },
       expect.objectContaining({
         brick: sourceElem,