+++ packages/brick-dll/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.31](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@1.3.30...@easyops/brick-dll@1.3.31) (2020-08-27)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [1.3.30](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@1.3.29...@easyops/brick-dll@1.3.30) (2020-08-26)
+
+### Bug Fixes
+
+- make dll manifest not changed ([6f58536](https://git.easyops.local/anyclouds/next-core/commits/6f58536))
+
 # [2.0.0-next.3](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@1.3.29...@easyops/brick-dll@2.0.0-next.3) (2020-08-25)
 
 ### Features