+++ packages/brick-types/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.52.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-types@1.51.0...@easyops/brick-types@1.52.0) (2020-08-26)
+
+### Features
+
+- 使用 channel 为 key 实现订阅消息分发 ([2e09abd](https://git.easyops.local/anyclouds/next-core/commits/2e09abd))
+
 # [2.0.0-next.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-types@1.51.0...@easyops/brick-types@2.0.0-next.2) (2020-08-25)
 
 ### Features