+++ packages/build-config-factory/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.18.9](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.8...@easyops/build-config-factory@1.18.9) (2020-08-26)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
 # [2.0.0-next.3](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.8...@easyops/build-config-factory@2.0.0-next.3) (2020-08-25)
 
 **Note:** Version bump only for package @easyops/build-config-factory